Move display collection tool to own route, renamed to builder
