+++ src/components/Builder/BuilderTool.tsx
@@ -1,11 +1,11 @@
-import { SelectionContainer } from 'components/Tools/Tools/Structure/CollectionOverview/Selection/SelectionContainer/SelectionContainer';
-import { CollectionContainer } from 'components/Tools/Tools/Structure/CollectionOverview/Collection/CollectionContainer/CollectionContainer';
+import { SelectionContainer } from 'components/Builder/Selection/SelectionContainer/SelectionContainer';
+import { CollectionContainer } from 'components/Builder/Collection/CollectionContainer/CollectionContainer';
 import type { RootState } from 'src/config/redux/store'
 import { useSelector, useDispatch } from 'react-redux'
 import { setEquipment, resetEquipment } from 'src/config/redux/slices/equipmentDisplaySlice'
 import { setCollection, resetCollection } from 'src/config/redux/slices/collectionDisplaySlice'
 import { useEffect, useState } from 'react';
-import { StringTextField } from 'components/Tools/Tools/Structure/CollectionOverview/StringTextField/StringTextField';
+import { StringTextField } from 'components/Builder/StringTextField/StringTextField';
 import { collectionInitialState } from 'src/config/redux/slices/collectionDisplaySlice';
 import { equipmentInitialState } from 'src/config/redux/slices/equipmentDisplaySlice';
 import { useSearchParams } from 'react-router-dom';
@@ -14,7 +14,7 @@ import { saveBuild, getBuild } from 'src/config/api/services/buildUrl';
 
 
 
-export const CollectionDisplay = () => {
+export const BuilderTool = () => {
   const dispatch = useDispatch();
   const collection = useSelector((state: RootState) => state.collectionDisplay)
   const equipment = useSelector((state: RootState) => state.equipmentDisplay)
@@ -32,16 +32,16 @@ export const CollectionDisplay = () => {
       return;
     }
     const decodedCollectionString = atob(response.data.build)
-      const [collectionString, equipmentString] = decodedCollectionString.split('|')
-      const collectionData = JSON.parse(collectionString)
-      const equipmentData = JSON.parse(equipmentString)
-      if (collectionData) {
-        dispatch(setCollection(collectionData))
-      }
-      if (equipmentData) {
-        dispatch(setEquipment(equipmentData))
-      }
-      setSearchParams({})
+    const [collectionString, equipmentString] = decodedCollectionString.split('|')
+    const collectionData = JSON.parse(collectionString)
+    const equipmentData = JSON.parse(equipmentString)
+    if (collectionData) {
+      dispatch(setCollection(collectionData))
+    }
+    if (equipmentData) {
+      dispatch(setEquipment(equipmentData))
+    }
+    setSearchParams({})
 
   }
 
@@ -118,9 +118,9 @@ export const CollectionDisplay = () => {
     <>
       <PopupModal isOpen={popupModal} onClose={() => { setPopupModal(false) }}>
         <div className='p-4 min-w-96'>
-        Share string:
-        <StringTextField>{shareString}</StringTextField>
-        Url:
+          Share string:
+          <StringTextField>{shareString}</StringTextField>
+          Url:
           <StringTextField>{shareUrl}</StringTextField>
         </div>
       </PopupModal>
