+++ src/components/Builder/EquippedContainer/EquippedContainer.tsx
